[DEV-4069] Restructure abstract classes
