+++ usaspending_api/search/tests/integration/spending_by_category/test_elasticsearch_header.py
@@ -18,7 +18,7 @@ def test_elasticsearch_headers(client, monkeypatch, elasticsearch_transaction_in
 def elasticsearch_http_header_helper(client, monkeypatch, elasticsearch_transaction_index, endpoint_name):
     logging_statements = []
     monkeypatch.setattr(
-        "usaspending_api.search.v2.views.spending_by_category_views.abstract_spending_by_category.logger.info",
+        "usaspending_api.search.v2.views.spending_by_category_views.spending_by_category.logger.info",
         lambda message: logging_statements.append(message),
     )
     monkeypatch.setattr(
