+++ usaspending_api/search/tests/integration/spending_by_category/utilities.py
@@ -3,7 +3,7 @@ from django.conf import settings
 
 def setup_elasticsearch_test(monkeypatch, elasticsearch_transaction_index, logging_statements):
     monkeypatch.setattr(
-        "usaspending_api.search.v2.views.spending_by_category_views.abstract_spending_by_category.logger.info",
+        "usaspending_api.search.v2.views.spending_by_category_views.spending_by_category.logger.info",
         lambda message: logging_statements.append(message),
     )
     monkeypatch.setattr(
